+++ tests/Core.Tests/Core.Tests.csproj
@@ -186,7 +186,7 @@
   <ItemGroup>
     <Compile Include="Admin\HealthTests.cs" />
     <Compile Include="Admin\IndexTests.cs" />
-    <Compile Include="Extensions\EnumExtensionsTests.cs" />
+    <Compile Include="Extensions\SimpleSearchExtensionsTests.cs" />
     <Compile Include="ServerInfoTests.cs" />
     <Compile Include="CoreIndexerTests.cs" />
     <Compile Include="BulkOperationTests.cs" />
